Move sync queue implementation to sync service folder
Not really cron specific in implementation.
